+++ packages/cli-forger/src/workers/folders/getPath.ts
@@ -1,0 +1,8 @@
+import { isArray } from 'lodash'
+import path from 'path'
+
+export function getPath(rawPath: string[] | string) {
+  const rawPaths = isArray(rawPath) ? rawPath : [rawPath]
+  const pathToCheck = path.resolve(...rawPaths)
+  return pathToCheck
+}