+++ packages/cli-forger/src/workers/logger/error.ts
@@ -1,0 +1,9 @@
+import chalk from 'chalk'
+import { workers } from '../index'
+
+export function error(...args: unknown[]) {
+  workers.logger.log(
+    chalk.bgRedBright.white.bold('[ERROR]'),
+    ...args.map((arg) => chalk.red(arg)),
+  )
+}