+++ packages/cli-forger/src/workers/logger/index.ts
@@ -1,0 +1,4 @@
+export { error } from './error'
+export { log } from './log'
+export { info } from './info'
+export * as exit from './exit'