+++ packages/eslint-config/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @rubykgen/eslint-config
+# @vortecx/eslint-config
+
+## 0.0.5
+
+### Patch Changes
+
+- Adding packages
 
 ## 0.0.4