+++ packages/ts-config/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @rubykgen/ts-config
+# @vortecx/ts-config
+
+## 0.0.3
+
+### Patch Changes
+
+- Adding packages
 
 ## 0.0.2